refactor: library constants to own file
