+++ packages/cli/src/formats/files/collectFiles.ts
@@ -10,7 +10,7 @@ import { aggregateInlineTranslations } from '../../translation/stage.js';
 import type { FileToUpload, JsxChildren } from 'generaltranslation/types';
 import { hashStringSync } from '../../utils/hash.js';
 import { TEMPLATE_FILE_NAME, TEMPLATE_FILE_ID } from '../../utils/constants.js';
-import { isInlineLibrary } from '../../react/jsx/utils/constants.js';
+import { isInlineLibrary } from '../../types/libraries.js';
 
 export async function collectFiles(
   options: TranslateFlags,
