+++ packages/cli/src/react/parse/wrapContent.ts
@@ -20,7 +20,7 @@ import {
 } from '../jsx/utils/parseAst.js';
 import { DEFAULT_SRC_PATTERNS } from '../../config/generateSettings.js';
 import { matchFiles } from '../../fs/matchFiles.js';
-import { Libraries } from '../jsx/utils/constants.js';
+import { Libraries } from '../../types/libraries.js';
 
 const IMPORT_MAP = {
   T: { name: 'T', source: Libraries.GT_REACT },
